💄Design: add check icon on vote page
