+++ src/styles/pages/hanging.module.scss
@@ -29,7 +29,7 @@
     $size: 40%;
     width: $size;
     height: calc($size + 30%);
-    background-image: url('../../assets/images/hanging_tree.svg');
+    background-image: url('@assets/images/hanging_tree.svg');
     background-repeat: no-repeat;
     background-size: contain;
   }
@@ -38,7 +38,7 @@
     position: absolute;
     height: 100%;
     width: 100%;
-    background-image: url('../../assets/images/hanging_leaves.svg');
+    background-image: url('@assets/images/hanging_leaves.svg');
     background-repeat: no-repeat;
     background-size: contain;
   }
